backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai-whisper==20231117
yt-dlp==2024.3.10
transformers==4.36.0
torch==2.1.0
sentencepiece==0.1.99
protobuf==4.25.1

